requirements.txt
aiofiles==0.4.0
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
bcrypt==3.1.6
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
httptools==0.0.13
idna==2.8
idna-ssl==1.1.0
jsonschema==3.0.1
motor==2.0.0
multidict==4.5.2
passlib==1.7.1
pycparser==2.19
PyJWT==1.7.1
pymongo==3.8.0
pyrsistent==0.15.2
python-dateutil==2.8.0
python-slugify==3.0.2
raven==6.10.0
raven-aiohttp==0.7.0
requests==2.22.0
sanic==19.3.1
sanic-mongo==1.7.1
sanic-sentry-error-handler==0.1.7
sentry-sdk==0.10.2
six==1.12.0
text-unidecode==1.2
typing-extensions==3.7.4
ua-parser==0.8.0
ujson==1.35
urllib3==1.25.3
user-agents==2.0
uvloop==0.12.2
websockets==6.0
yarl==1.3.0
geoip2